refactor: extract versions from lockfile via the Docker build
Rather than running commands outside the build to determine dependency
versions, we now have a bake target that does this. This avoids the
problem of having to install dev dependencies outside the build, and
also means we pull everything from the lockfile, which means we don't
need to do any post-processing of dependabot PRs that just change the
lockfile.